+++ test/random.js
@@ -11,7 +11,7 @@ describe('Weird and random suite', function() {
     it('Setup mysql-cache', done => {
         db = new MysqlCache(settings)
 
-        db.event.on('connected', () => {
+        db.connectAsync().then(() => {
             done()
         })
     })
@@ -23,29 +23,15 @@ describe('Weird and random suite', function() {
 
     it('test trace', () => {
         settings.verbose = true
-        db.util.trace('kappa')
+        db.trace('kappa')
     })
 
     it('test error', () => {
-        db.util.error(new Error('TESTTT :D'))
+        db.error(new Error('TESTTT :D'))
     })
 
     it('test error #2', () => {
-        settings.prettyError = false
-        settings.printErrors = true
-        db.util.error(new Error('TESTTT :D'))
-    })
-
-    it('test FATAL error', () => {
-        settings.prettyError = false
-        settings.printErrors = true
-        const klappaError = new Error('TESTTT :D')
-
-        klappaError.fatal = true
-
-        assert.throws(() => {
-            db.util.error(klappaError)
-        }, Error)
+        db.error(new Error('TESTTT :D'))
     })
 
     it('cacheprovider: illigal run action', () => {
@@ -58,60 +44,39 @@ describe('Weird and random suite', function() {
         }, Error)
     })
 
-    it('cacheprovider: illigal cacheprovider action', done => {
+    it('cacheprovider: illigal cacheprovider action', () => {
         settings.cacheProvider = 'clappa'
-        settings.prettyError = true
-        settings.printErrors = true
-        const newClappaDB = new MysqlCache(settings)
-
-        newClappaDB.event.on('connected', () => {
-            newClappaDB.query({sql:'select 1 + 1 as solution'}, (err, resultMysql, mysqlCache) => {
-                if (err) {
-                    throw new Error(err)
-                }
-                done()
-            })
-        })
-    })
-
-    it('cacheprovider: run without callback', done => {
-        settings.cacheProvider = 'lru'
-        settings.prettyError = true
-        settings.printErrors = true
-        const newClappaDB = new MysqlCache(settings)
-
-        newClappaDB.event.on('connected', () => {
-            newClappaDB.cacheProvider.run('get')
-            done()
-        })
+        assert.throws(() => {
+            const db = new MysqlCache(settings)
+        }, Error)
     })
 
     it('flush without cb', done => {
         settings.cacheProvider = 'lru'
         settings.prettyError = true
         settings.printErrors = true
-        const newClappaDB = new MysqlCache(settings)
+        const db = new MysqlCache(settings)
 
-        newClappaDB.event.on('connected', () => {
-            newClappaDB.flush()
+        db.connectAsync().then(() => {
+            db.flush()
             done()
         })
     })
 
     it('end undefined pool', done => {
-        const newClappaDB = new MysqlCache(settings)
+        const db = new MysqlCache(settings)
 
-        newClappaDB.event.on('connected', () => {
-            assert.equal(newClappaDB.endPool(undefined), false)
+        db.connectAsync().then(() => {
+            assert.equal(db.endPool(undefined), false)
             done()
         })
     })
 
     it('query error', done => {
-        const newClappaDB = new MysqlCache(settings)
+        const db = new MysqlCache(settings)
 
-        newClappaDB.event.on('connected', () => {
-            newClappaDB.query({sql:'select 1 + 1 asfsdfdsfsd dfs 432134  tgfd solution'}, (err, resultMysql, mysqlCache) => {
+        db.connectAsync().then(() => {
+            db.query({sql:'select 1 + 1 asfsdfdsfsd dfs 432134  tgfd solution'}, (err, resultMysql, mysqlCache) => {
                 assert.throws(() => {
                     if (err) {
                         throw new Error(err)